+++ lib/object_inspector/formatters/base_formatter.rb
@@ -1,9 +1,8 @@
 # frozen_string_literal: true
 
-# ObjectInspector::BaseFormatter is an abstract base class that interfaces
-# with {ObjectInspector::Inspector} objects to combine the supplied
-# {#identification}, {#flags}, {#info}, and {#name} strings into a friendly
-# "inspect" String.
+# An abstract base class that interfaces with {ObjectInspector::Inspector}
+# objects to combine the supplied {#identification}, {#flags}, {#info}, and
+# {#name} strings into a friendly "inspect" String.
 class ObjectInspector::BaseFormatter
   attr_reader :inspector
 
@@ -21,8 +20,8 @@ class ObjectInspector::BaseFormatter
 
   # Delegates to {Inspector#wrapped_object_inspection_result}.
   #
-  # @return [String] if given
-  # @return [NilClass] if not given
+  # @return [String] If given.
+  # @return [NilClass] If not given.
   def wrapped_object_inspection_result
     @wrapped_object_inspection_result ||=
       @inspector.wrapped_object_inspection_result
@@ -30,39 +29,39 @@ class ObjectInspector::BaseFormatter
 
   # Delegates to {Inspector#identification}.
   #
-  # @return [String] if given
+  # @return [String] If given.
   def identification
     @identification ||= @inspector.identification
   end
 
   # Delegates to {Inspector#flags}.
   #
-  # @return [String] if given
-  # @return [NilClass] if not given
+  # @return [String] If given.
+  # @return [NilClass] If not given.
   def flags
     @flags ||= @inspector.flags
   end
 
   # Delegates to {Inspector#issues}.
   #
-  # @return [String] if given
-  # @return [NilClass] if not given
+  # @return [String] If given.
+  # @return [NilClass] If not given.
   def issues
     @issues ||= @inspector.issues
   end
 
   # Delegates to {Inspector#info}.
   #
-  # @return [String] if given
-  # @return [NilClass] if not given
+  # @return [String] If given.
+  # @return [NilClass] If not given.
   def info
     @info ||= @inspector.info
   end
 
   # Delegates to {Inspector#name}.
   #
-  # @return [String] if given
-  # @return [NilClass] if not given
+  # @return [String] If given.
+  # @return [NilClass] If not given.
   def name
     @name ||= @inspector.name
   end