+++ test/object_inspector/interrogate_object_test.rb
@@ -2,23 +2,19 @@
 
 require "test_helper"
 
-class ObjectInspector::ObjectInterrogatorTest < Minitest::Spec
-  let(:unit_class) { ObjectInspector::ObjectInterrogator }
+class ObjectInspector::InterrogateObjectTest < Minitest::Spec
+  describe ".call" do
+    subject { ObjectInspector::InterrogateObject }
 
-  describe "#call" do
     given "an Object with no keyword Arguments" do
-      subject {
-        unit_class.new(
-          object: SimpleTestObject.new,
-          method_name:,
-        )
-      }
+      let(:object) { SimpleTestObject.new }
 
       given "a public method_name" do
         let(:method_name) { :example_method }
 
         it "returns the expected result" do
-          _(subject.call).must_equal("a")
+          result = subject.call(object, method_name:)
+          _(result).must_equal("a")
         end
       end
 
@@ -26,7 +22,8 @@ class ObjectInspector::ObjectInterrogatorTest < Minitest::Spec
         let(:method_name) { :example_private_method }
 
         it "returns the expected result" do
-          _(subject.call).must_equal("b")
+          result = subject.call(object, method_name:)
+          _(result).must_equal("b")
         end
       end
 
@@ -34,25 +31,22 @@ class ObjectInspector::ObjectInterrogatorTest < Minitest::Spec
         let(:method_name) { :invalid_method }
 
         it "returns nil" do
-          _(subject.call).must_be_nil
+          result = subject.call(object, method_name:)
+          _(result).must_be_nil
         end
       end
     end
 
     given "an Object with all optional keyword Arguments" do
-      subject {
-        unit_class.new(
-          object: AllOptionalKeywordArgumentsTestObject.new,
-          method_name: :example_method,
-          kwargs:,
-        )
-      }
+      let(:object) { AllOptionalKeywordArgumentsTestObject.new }
+      let(:method_name) { :example_method }
 
       given "a valid keyword argument" do
         let(:kwargs) { { a: "a" } }
 
         it "returns the expected result" do
-          _(subject.call).must_equal(["a", 2])
+          result = subject.call(object, method_name:, kwargs:)
+          _(result).must_equal(["a", 2])
         end
       end
 
@@ -60,7 +54,8 @@ class ObjectInspector::ObjectInterrogatorTest < Minitest::Spec
         let(:kwargs) { { invalid_kwarg: 1 } }
 
         it "returns the expected result" do
-          _(subject.call).must_equal([1, 2])
+          result = subject.call(object, method_name:, kwargs:)
+          _(result).must_equal([1, 2])
         end
       end
 
@@ -68,43 +63,38 @@ class ObjectInspector::ObjectInterrogatorTest < Minitest::Spec
         let(:kwargs) { {} }
 
         it "returns the expected result" do
-          _(subject.call).must_equal([1, 2])
+          result = subject.call(object, method_name:, kwargs:)
+          _(result).must_equal([1, 2])
         end
       end
     end
 
     given "an Object with some optional keyword Arguments" do
-      subject {
-        unit_class.new(
-          object: SomeOptionalKeywordArgumentsTestObject.new,
-          method_name: :example_method,
-          kwargs:,
-        )
-      }
+      let(:object) { SomeOptionalKeywordArgumentsTestObject.new }
+      let(:method_name) { :example_method }
 
       given "a missing required keyword argument" do
         let(:kwargs) { { b: "b" } }
 
         it "raises ArgumentError" do
-          _(-> { subject.call }).must_raise(ArgumentError)
+          _(-> {
+            subject.call(object, method_name:, kwargs:)
+          }).must_raise(ArgumentError)
         end
       end
     end
 
     given "an Object with all required keyword Arguments" do
-      subject {
-        unit_class.new(
-          object: AllRequiredKeywordArgumentsTestObject.new,
-          method_name: :example_method,
-          kwargs:,
-        )
-      }
+      let(:object) { AllRequiredKeywordArgumentsTestObject.new }
+      let(:method_name) { :example_method }
 
       given "no keyword arguments" do
         let(:kwargs) { {} }
 
         it "raises ArgumentError" do
-          _(-> { subject.call }).must_raise(ArgumentError)
+          _(-> {
+            subject.call(object, method_name:, kwargs:)
+          }).must_raise(ArgumentError)
         end
       end
     end