+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+### 1.0.0 - 2025-10-12
+
 #### BREAKING CHANGES
 
 - Refactor `ObjectInspector::InspectorsHelper` -> `ObjectInspector::Behaviors`